Temp fix for the viewer not opening files readable by the old editor.

+++ rpfm_ui/src/packedfile_views/rigidmodel/mod.rs
@@ -78,16 +78,20 @@ impl PackedFileRigidModelView {
             },
 
             #[cfg(feature = "support_model_renderer")] renderer: {
-                let renderer = create_q_rendering_widget(&mut file_view.main_widget().as_ptr())?;
+                if let Ok(renderer) = create_q_rendering_widget(&mut file_view.main_widget().as_ptr()) {
 
-                // We need to manually kill the renderer or it'll keep lagging the UI.
-                if let Err(error) = add_new_primary_asset(&renderer.as_ptr(), &file_view.path().read().unwrap(), data.data()) {
-                    renderer.delete();
-                    return Err(error);
+                    // We need to manually kill the renderer or it'll keep lagging the UI.
+                    if let Err(error) = add_new_primary_asset(&renderer.as_ptr(), &file_view.path().read().unwrap(), data.data()) {
+                        renderer.delete();
+
+                        QWidget::new_1a(file_view.main_widget())
+                    } else {
+                        splitter.add_widget(&renderer);
+                        renderer
+                    }
+                } else {
+                    QWidget::new_1a(file_view.main_widget())
                 }
-
-                splitter.add_widget(&renderer);
-                renderer
             },
 
             #[cfg(feature = "support_model_renderer")] path: Some(file_view.path_raw()),
@@ -118,7 +122,7 @@ impl PackedFileRigidModelView {
 
         #[cfg(feature = "support_model_renderer")] {
             if let Some(ref path) = self.path {
-                add_new_primary_asset(&self.renderer.as_ptr(), &path.read().unwrap(), data.data())?;
+                let _ = add_new_primary_asset(&self.renderer.as_ptr(), &path.read().unwrap(), data.data());
             }
         }
